color is totally in linear spaces now. basecolor stored in the disk is in sRGB space, because this makes color rich.

+++ Source/Vertex.h
@@ -8,6 +8,7 @@
 
 #pragma once
 #include "MathUtil.h"
+#include "half.h"
 #include "Platform.h"
 #include <memory>
 namespace JustRay
@@ -18,8 +19,8 @@ struct Vertex
     Int_2_10_10_10 normal;
     Int_2_10_10_10 tangent;
     Int_2_10_10_10 bitangent;
-    unsigned short tx;
-    unsigned short ty;
+    half tx;
+    half ty;
     bool operator<(const Vertex& rhs) const;
 };
 struct VertexAttributeDescription
